requirements.txt
# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Configuration
python-dotenv>=1.0.0

# LLM providers
openai>=1.0.0
ollama>=0.1.0

# Vector database & embeddings
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
torch>=2.1.0
tensorflow>=2.16.0
tf-keras

# Datasets
datasets==3.6.0

# Progress bars
tqdm>=4.66.0

# Testing
pytest>=7.4.0

# RSS + Web scraping
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.1
